requirements.txt
gymnasium
torch
numpy
opencv-python
imageio
pillow